Basic initialization method for arc binning functions.
